+++ cadastre/classes/atlasPrint.listener.php
@@ -64,6 +64,7 @@ class atlasPrintListener extends jEventListener
                 $p,
                 array(
                     'service' => 'CADASTRE',
+                    'version' => '1.0.0',
                     'request' => 'getHtml',
                     'layer' => $layer,
                     'parcelle' => $geo_parcelle,